+++ Sources/Definition.swift
@@ -34,12 +34,14 @@ public struct DefinitionKey: Hashable, CustomStringConvertible {
     self.tag = tag
   }
   
-  public var hashValue: Int {
-    return "\(type)-\(typeOfArguments)-\(tag.desc)".hashValue
+  public func hash(into hasher: inout Hasher) {
+    hasher.combine(ObjectIdentifier(type))
+    hasher.combine(ObjectIdentifier(typeOfArguments))
+    hasher.combine(tag.desc)
   }
   
   public var description: String {
-    return "type: \(type), arguments: \(typeOfArguments), tag: \(tag.desc)"
+    return "type: \(String(reflecting: type)), arguments: \(typeOfArguments), tag: \(tag.desc)"
   }
   
   func tagged(with tag: DependencyContainer.Tag?) -> DefinitionKey {
@@ -53,7 +55,7 @@ public struct DefinitionKey: Hashable, CustomStringConvertible {
     return
       lhs.type == rhs.type &&
       lhs.typeOfArguments == rhs.typeOfArguments &&
-      lhs.tag == rhs.tag
+      lhs.tag.desc == rhs.tag.desc
   }
   
 }
@@ -124,6 +126,22 @@ public final class Definition<T, U>: DefinitionType {
       self.resolveProperties = { try block($0, $1 as! T) }
     }
     return self
+  }
+
+  @discardableResult public func resolvingProperty<Root, V>(_ keyPath: ReferenceWritableKeyPath<Root, V>, as type: Any.Type = V.self, tag: DependencyTagConvertible? = nil) -> Definition {
+    return resolvingProperties { (container, instance) in
+      precondition(instance is Root, "Type of resolved instance \(Swift.type(of: instance)) does not match expected type \(Root.self)")
+      let resolved = try container.resolve(type, tag: tag)
+      precondition(resolved is V, "Type of resolved property \(Swift.type(of: resolved)) does not match expected type \(type)")
+      (instance as! Root)[keyPath: keyPath] = resolved as! V
+    }
+  }
+
+  @discardableResult public func resolvingProperty<Root, V>(_ keyPath: ReferenceWritableKeyPath<Root, V>, factory: @escaping (DependencyContainer) throws -> V = { try $0.resolve() }) -> Definition {
+    return resolvingProperties { (container, instance) in
+      precondition(instance is Root, "Type of resolved instance \(Swift.type(of: instance)) does not match expected type \(Root.self)")
+      (instance as! Root)[keyPath: keyPath] = try factory(container)
+    }
   }
 
   /**