+++ Sources/Utils.swift
@@ -56,6 +56,13 @@ class Box<T> {
   }
 }
 
+class NullableBox<T> {
+    var unboxed: T?
+    init(_ value: T?) {
+        self.unboxed = value
+    }
+}
+
 protocol WeakBoxType {
   var unboxed: AnyObject? { get }
 }
@@ -66,15 +73,8 @@ class WeakBox<T>: WeakBoxType {
     return unboxed as? T
   }
 
-  init(_ value: T) {
-    #if _runtime(_ObjC)
-      weak var value: AnyObject? = value as AnyObject
-    #else
-      weak var value: AnyObject? = value as? AnyObject
-    #endif
-    guard value != nil else {
-      fatalError("Can not store weak reference to not a class instance (\(T.self))")
-    }
+  init(_ value: T?) {
+    weak var value: AnyObject? = value as AnyObject
     self.unboxed = value
   }
 }